Divide Problem.cpp and Problem.hpp in 3 separate files, one for each applicative case
